// SDKs/twilio.com/twilio_conversations_v1/1.37.3/typescript/src/sdk/models/shared/conversationsv1credential.ts
import { SpeakeasyMetadata, SpeakeasyBase } from "../../../internal/utils";
import { CredentialEnumPushTypeEnum } from "./credentialenumpushtypeenum";



export class ConversationsV1Credential extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "json, name=account_sid" })
  accountSid?: string;

  @SpeakeasyMetadata({ data: "json, name=date_created" })
  dateCreated?: Date;

  @SpeakeasyMetadata({ data: "json, name=date_updated" })
  dateUpdated?: Date;

  @SpeakeasyMetadata({ data: "json, name=friendly_name" })
  friendlyName?: string;

  @SpeakeasyMetadata({ data: "json, name=sandbox" })
  sandbox?: string;

  @SpeakeasyMetadata({ data: "json, name=sid" })
  sid?: string;

  @SpeakeasyMetadata({ data: "json, name=type" })
  type?: CredentialEnumPushTypeEnum;

  @SpeakeasyMetadata({ data: "json, name=url" })
  url?: string;
}